--- client/src/helpers/history.js
@@ -1,3 +1,0 @@
-// import { createBrowserHistory } from 'history';
-
-// export const history = createBrowserHistory();
